+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - feat: remove `Triangle2` and `Triangle3` types, passing individual Vec2/Vec3 arguments has been generally more practical instead
 - feat: add `quat.fromDegrees` convenience function to set quaternion from Euler angles in degrees
 - feat: add `box3.union` to compute the union of two Box3s
+- feat: add `box3.size` to compute the size or dimensions of a Box3
 
 ## 0.0.6